+++ src/day_01/part_2.rs
@@ -2,6 +2,6 @@
 const INPUT: &str = include_str!("input.txt");
 // const INPUT: &str = include_str!("small.txt");
 
-pub fn run() -> Result<i64, String> {
+pub fn run() -> Result<usize, String> {
     todo!()
 }
